Add files via upload
#Another FIX
Error: Template validation error: Template format error: Unresolved resource dependencies [Secondary1NodeSubnet] in the Resources block of the template

When someone tries to Create a stack with ClusterShardCount > 0 and ClusterReplicaSetCount=1.

The Secondary1NodeSubnet condition was to be created when ClusterReplicaSetCount is set to 3. 


Fixed the conditions

        "CreateThreeReplicaSet": {
   "Fn::Or" : [
      {"Fn::Equals": [{ "Ref": "ClusterReplicaSetCount"}, "3"]},
      {"Fn::Equals": [{ "Ref": "ClusterShardCount"}, "1"]},
      {"Fn::Equals": [{ "Ref": "ClusterShardCount"}, "2"]},
      {"Fn::Equals": [{ "Ref": "ClusterShardCount"}, "3"]},
   ]
        },
 

Important Note: 

You can not upgrade the ClusterShardCount after the creation of the stack. If someone will try to update the stack after modifying ClusterShardCount value, that will fail the PrimaryReplicaNode20WaitForNodeInstallGP2.

There is no fix for that.